Modulo 3 Completo

+++ app/Http/Controllers/Admin/PlanController.php
@@ -52,9 +52,17 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('url', $url)->first();
         if (!$plan) {
             return redirect()->back();
+        }
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'Existem  detalhes vinculados a este plano, portanto não é possivel remover');
         }
 
         $plan->delete();
